engine: the last of the smaller correctness notes, and four costs (F, L)
**Six notes that each named a check standing in the wrong place.**

`InputSample::is_finite` was not the gate its callers thought it was. Arc length is a
*difference* of two positions, and two finite reports a whole `f32` range apart differ
by infinity — from which the fitter divides `inf / inf` and hands `spline`'s solve the
NaN its `unreachable!` says cannot arrive. So the gate bounds the position as well:
`is_admissible`, against `COORD_LIMIT`, which is the first coordinate the `i32` tile
grid cannot address rather than a number picked to be large. Exclusive, because
`TileRect::covering` refuses that value itself. The fitter's poison table grows the two
rows that separate the new gate from the old one, and the boundary row fails against
`<=`.

**The readback owns the claim about the texture it decodes.** `read_many_rgba16f`
decodes four halves a texel; that is a fact about the texture, and it was asserted at
its two call sites in `debug_assert`s a release build did not run. It is one
unconditional check inside the reader now. Nothing outside the process picks these
formats — they come from the color space this build compiled — so an assert is the
right shape and there is no `Result` arm to invent.

`plan_invert` named its tiles in `rpds`' hasher order where every other planner names
them in the row-major order `TileRect::coords` walks. No pixel moved either way — each
tile is an independent clear-and-draw — but an invariant that holds everywhere is worth
the sort on a set bounded by `MAX_SELECTION_TILES`.

`Resource::Existence` silently also means the layer's *minted* kind, which
`cannot_carry` and `set_layer_blend` read without declaring: sound, because no action
converts one, and now said where a future "convert to matte" will read it. Being a
*group* is deliberately not on that list — a leaf becomes one under `MoveLayer`, so
what covers it is `StackOrder`. `PlaceImage`'s out-of-log store is named in the same
place, since that enum is where a reader asks what determinism rests on.
`start_collaboration` says that a scrubbed-away future does not survive the share.

**Four costs that scale with the wrong thing (L).** A substrate's PNG was decoded
again for every scale it is laid at: the registry now decodes once, at registration,
and keeps whatever the resource says is worth keeping — the substrate's height field,
and deliberately nothing for an environment, whose decode is megabytes read once by a
mip chain. That door also subsumes `Environment::decodes`, so validation is one place
rather than a call each caller had to remember.

`SelectionRenderer::constant` built and uploaded a texture per call for a partial
coverage, which `mask_for` asks once per tile of every fill, transform and stroke under
an inverted-partial selection; 256 slots of one texel cover every byte it can be asked
for. The tile pool's per-format free lists are scanned rather than hashed — four of
them, on the hottest path in the crate. The scratch pool's label is out of the match:
identical squares under two names kept two free lists and held half as much warm.

The pen-orientation bake — two 64 MiB buffers and 16 M `ln` calls inside the store
lock, mid-gesture — now opens `asset.pen_bake`, and both askers reach it through one
method rather than two copies of the lazy bake.

Master was red on three counts and is not now: two clippy warnings and a `rustfmt`
diff, none of them mine.

+++ crates/stark-ui/src/platform.rs
@@ -1492,7 +1492,7 @@ pub async fn normalize_substrate_image(bytes: Vec<u8>) -> Result<Vec<u8>, String
     // deep trench around a canvas that has none. White, not black, because an
     // unpainted substrate reads as its own top surface.
     for p in px.as_chunks_mut::<4>().0 {
-        let lum = ((77 * p[0] as u32 + 150 * p[1] as u32 + 29 * p[2] as u32) >> 8) as u32;
+        let lum = (77 * p[0] as u32 + 150 * p[1] as u32 + 29 * p[2] as u32) >> 8;
         let a = p[3] as u32;
         let over = (lum * a + 255 * (255 - a)) / 255;
         let g = over as u8;
